NatsSub redesign
