+++ src/NATS.Client.JetStream/INatsJSSubConsume.cs
@@ -1,0 +1,8 @@
+using System.Threading.Channels;
+
+namespace NATS.Client.JetStream;
+
+public interface INatsJSSubConsume<T> : IAsyncDisposable
+{
+    ChannelReader<NatsJSMsg<T?>> Msgs { get; }
+}
