+++ tests/NATS.Client.JetStream.Tests/JetStreamTest.cs
@@ -83,9 +83,10 @@ public class JetStreamTest
             // Consume
             var cts2 = new CancellationTokenSource(TimeSpan.FromSeconds(10));
             var messages = new List<NatsJSMsg<TestData?>>();
-            await foreach (var msg in consumer.ConsumeAsync<TestData>(
-                               new NatsJSConsumeOpts { MaxMsgs = 100 },
-                               cancellationToken: cts2.Token))
+            var cc = await consumer.ConsumeAsync<TestData>(
+                new NatsJSConsumeOpts { MaxMsgs = 100 },
+                cancellationToken: cts2.Token);
+            await foreach (var msg in cc.Msgs.ReadAllAsync(cts2.Token))
             {
                 messages.Add(msg);
 
@@ -104,11 +105,12 @@ public class JetStreamTest
             Assert.Equal(2, messages.Count);
             Assert.Equal("events.foo", messages[0].Msg.Subject);
             Assert.Equal("events.foo", messages[1].Msg.Subject);
+            var cc2 = await consumer.ConsumeAsync<TestData>(
+                new NatsJSConsumeOpts { MaxMsgs = 100 },
+                cancellationToken: cts2.Token);
 
             // Consume the unacknowledged message
-            await foreach (var msg in consumer.ConsumeAsync<TestData>(
-                               new NatsJSConsumeOpts { MaxMsgs = 100 },
-                               cancellationToken: cts2.Token))
+            await foreach (var msg in cc2.Msgs.ReadAllAsync(cts2.Token))
             {
                 Assert.Equal("events.foo", msg.Msg.Subject);
                 break;
